+++ app/api/planting_data.py
@@ -4,7 +4,7 @@ from flask import request, jsonify
 from flask_openapi3 import Tag, APIBlueprint
 
 from app.config import API_PREFIX, API_VERSION
-from app.dto.api_responses import PlantingDataRecord, CropRecordResponse, Unauthorized
+from app.dto.planting_data_resp import PlantingDataRecord, CropRecordResponse, Unauthorized
 from app.dto.data_filters import PlantingDataFilter
 from app.repo.planting_data import PlantingDataRepo
 from app.utils.logging import SharedLogger
